hides progress buttons in edit mode
